requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-grpc
pytest-mock
pytest-aiohttp
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
twine>=4.0.0
build>=0.10.0
isort>=5.12.0
pytest-playwright-asyncio 